+++ src/lib/st-item-list/item-list-scroll/item-list-scroll.component.spec.ts
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { VirtualScrollModule } from 'angular2-virtual-scroll';
+import { VirtualScrollerModule } from 'ngx-virtual-scroller';
 import { times as _times } from 'lodash';
 
 // Components
@@ -39,7 +39,7 @@ let list: StItemListElement[] = generateData(10);
 describe('StItemListComponent', () => {
    beforeEach(async(() => {
       TestBed.configureTestingModule({
-         imports: [VirtualScrollModule],
+         imports: [VirtualScrollerModule],
          declarations: [ItemListScrollComponent, ItemListItemComponent]
       })
          .compileComponents();  // compile template and css