+++ src/lib/st-item-list/st-item-list.component.spec.ts
@@ -11,7 +11,7 @@
 import { DebugElement } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { VirtualScrollModule } from 'angular2-virtual-scroll';
+import { VirtualScrollerModule } from 'ngx-virtual-scroller';
 import { times as _times } from 'lodash';
 
 // Components
@@ -49,7 +49,7 @@ let list: StItemListElement[] = generateData(10);
 describe('StItemListComponent', () => {
    beforeEach(async(() => {
       TestBed.configureTestingModule({
-         imports: [StSearchModule, VirtualScrollModule],
+         imports: [StSearchModule, VirtualScrollerModule],
          declarations: [StItemListComponent, ItemListItemComponent, ItemListScrollComponent]
       })
          .compileComponents();  // compile template and css