+++ src/lib/st-two-list-selection/st-two-list-selection.component.spec.ts
@@ -12,7 +12,7 @@ import { DebugElement, SimpleChange, SimpleChanges, NO_ERRORS_SCHEMA } from '@an
 import { async, ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { Http } from '@angular/http';
 import { By } from '@angular/platform-browser';
-import { VirtualScrollModule } from 'angular2-virtual-scroll';
+import { VirtualScrollerModule } from 'ngx-virtual-scroller';
 import {
    times as _times,
    cloneDeep as _cloneDeep
@@ -54,7 +54,7 @@ let selectedElements: StTwoListSelectionElement[] = _cloneDeep(allElements.slice
 describe('StTwoListSelectionComponent', () => {
    beforeEach(async(() => {
       TestBed.configureTestingModule({
-         imports: [StSearchModule, PipesModule, VirtualScrollModule, StCheckboxModule],
+         imports: [StSearchModule, PipesModule, VirtualScrollerModule, StCheckboxModule],
          declarations: [StTwoListSelectionComponent, StTwoListSelectionViewComponent, ListSelectionComponent, ListItemComponent, ListScrollComponent],
          schemas: [NO_ERRORS_SCHEMA]
       })